split into precode_test and precode
