+++ precode_test.go
@@ -2,51 +2,11 @@ package main
 
 import (
     "net/http"
-    "strconv"
     "strings"
     "net/http/httptest"
     "testing"
     "github.com/stretchr/testify/assert"
 )
-
-var cafeList = map[string][]string{
-    "moscow": []string{"Мир кофе", "Сладкоежка", "Кофе и завтраки", "Сытый студент"},
-}
-
-func mainHandle(w http.ResponseWriter, req *http.Request) {
-    countStr := req.URL.Query().Get("count")
-    if countStr == "" {
-        w.WriteHeader(http.StatusBadRequest)
-        w.Write([]byte("count missing"))
-        return
-    }
-
-    count, err := strconv.Atoi(countStr)
-    if err != nil {
-        w.WriteHeader(http.StatusBadRequest)
-        w.Write([]byte("wrong count value"))
-        return
-    }
-
-    city := req.URL.Query().Get("city")
-
-    cafe, ok := cafeList[city]
-    if !ok {
-        w.WriteHeader(http.StatusBadRequest)
-        w.Write([]byte("wrong city value"))
-        return
-    }
-
-    if count > len(cafe) {
-        count = len(cafe)
-    }
-
-    answer := strings.Join(cafe[:count], ",")
-
-    w.WriteHeader(http.StatusOK)
-    w.Write([]byte(answer))
-}
-
 func TestMainHandlerStatusOk(t *testing.T) {
     req := httptest.NewRequest("GET", "/cafe?count=2&city=moscow", nil)
     responseRecorder := httptest.NewRecorder()
@@ -82,5 +42,3 @@ req := httptest.NewRequest("GET", "/cafe?count=2&city=omsk", nil)
     assert.Equal(t,"wrong city value", body, "Города нет в базе")
 
 } 
-func main() {
-}
